Move configureInPlace to localized usage
